BlackJack
The blackjack game is created using the Python programming language. It can be played on a laptop. The strategy for the game is straightforward and includes the following guidelines:

If the dealer's revealed card has a value of 4, 5, or 6, it might be advantageous to double your bet if you have an Ace and a 4 in your hand.

Consider surrendering if you have a total of 16 in your hand while the dealer's revealed card is a 9, 10, or Ace.

Always choose to split if you are dealt a pair of Aces.

If you receive a pair of 7s, only decide to hit (draw another card) if the dealer's revealed card is an 8, 9, 10, or Ace.

Feel free to play the blackjack game for free now and test whether the strategy proves effective!